Make discover function generic on SpecT.
Function discover now accepts and interprets two strings, pattern and
property.

discover function is now generic on the returning spec type:

oldDiscover ::
  ∀ m.
  MonadEffect m =>
  m (Spec Unit)

discover ::
  ∀ r m g i m'.
  MonadEffect m =>
  Applicative m' =>
  String ->
  String ->
  m (SpecT g i m' Unit)

It's up to the caller to know in advance the concrete types of discovered
specs.

A concrete example is shown in the test module Test.Spec.Discovery.Main